#30 Use redis for session storage; inline app config in prep for refac
