+++ api/src/authMiddleware.ts
@@ -2,6 +2,12 @@ import { Request, Response, NextFunction } from 'express';
 
 import { UnauthorizedError } from './httpErrors';
 
+declare module 'express-session' {
+  interface Session {
+    principalId: number;
+  }
+}
+
 export const loggedIn = (req: Request, res: Response, next: NextFunction) => {
   if (req.session.principalId) {
     next();
